Refactor and pull out QR code display
