+++ shared/address_explorer.py
@@ -75,7 +75,12 @@ async def show_n_addresses(path, addr_fmt, start, n):
     import version
 
     def make_msg(start):
-        msg = "Press 1 to save to MicroSD.\n\n"
+        msg = ''
+        if start == 0:
+            msg = "Press 1 to save to MicroSD."
+            if version.has_fatram:
+                msg += " 4 to view QR Codes."
+            msg += '\n\n'
         msg += "Addresses %d..%d:\n\n" % (start, start + n - 1)
 
         addrs = []
@@ -97,8 +102,6 @@ async def show_n_addresses(path, addr_fmt, start, n):
 
             stash.blank_object(node)
 
-        if version.has_fatram:
-            msg += "Press 4 to view QR Code. "
 
         msg += "Press 9 to see next group, 7 to go back. X to quit."
 
@@ -119,7 +122,8 @@ async def show_n_addresses(path, addr_fmt, start, n):
 
         if ch == '4':
             if not version.has_fatram: continue
-            await show_address_qr(addrs, (addr_fmt & AFC_BECH32), start)
+            from ux import show_qr_codes
+            await show_qr_codes(addrs, bool(addr_fmt & AFC_BECH32), start)
             continue
 
         if ch == '7' and start>0:
@@ -212,107 +216,5 @@ Press 4 to start.''', escape='4')
     path, addr_fmt = picked
 
     await show_n_addresses(path, addr_fmt, 0, 10)
-
-
-async def show_address_qr(addrs, is_segwit, start_n):
-    # show a QR code for the address. can only work on Mk3
-    # Version 2 would be nice, but can't hold what we need, even at min error correction,
-    # so we are forced into version 3 = 29x29 pixels
-    # - see <https://www.qrcode.com/en/about/version.html>
-    # - to display 29x29 pixels, we have to double them up: 58x58
-    # - not really providing enough space around it
-    # - inverted QR (black/white swap) still readable by scanners, altho wrong
-
-    from utils import imported
-    from display import FontSmall, FontTiny
-    import uQR as uqr
-    from main import dis
-
-    idx = 0             # start with first address
-    invert = False      # looks better, but neither mode is ideal
-
-    addr = addrs[idx]
-
-    def render(addr):
-        dis.busy_bar(True)
-        with imported('uQR') as uqr:
-            if is_segwit:
-                # targeting 'alpha numeric' mode, typical len is 42
-                ec = uqr.ERROR_CORRECT_Q
-                assert len(addr) <= 47
-            else:
-                # has to be 'binary' mode, altho shorter msg, typical 34-36
-                ec = uqr.ERROR_CORRECT_M
-                assert len(addr) <= 42
-
-            q = uqr.QRCode(version=3, box_size=1, border=0, mask_pattern=3, error_correction=ec)
-            if is_segwit:
-                here = uqr.QRData(addr.upper().encode('ascii'),
-                                        mode=uqr.MODE_ALPHA_NUM, check_data=False)
-            else:
-                here = uqr.QRData(addr.encode('ascii'), mode=uqr.MODE_8BIT_BYTE, check_data=False)
-            q.add_data(here)
-            q.make(fit=False)
-
-            return q.get_matrix()
-
-    data = render(addr)
-
-    def redraw():
-        dis.clear()
-
-        w = 29          # because version=3
-        XO,YO = 7, 3    # offsets
-
-        if not invert:
-            dis.dis.fill_rect(XO-YO, 0, 64, 64, 1)
-
-        for x in range(w):
-            for y in range(w):
-                px = data[x][y]
-                X = (x*2) + XO
-                Y = (y*2) + YO
-                dis.dis.fill_rect(X,Y, 2,2, px if invert else (not px))
-
-        x, y = 73, 0 if is_segwit else 2
-        ll = 7      # per line
-        for i in range(0, len(addr), ll):
-            dis.text(x, y, addr[i:i+ll], FontSmall)
-            y += 10 if is_segwit else 12
-
-        if not invert:
-            # show path number, very tiny
-            ai = str(start_n + idx)
-            if len(ai) == 1:
-                dis.text(0, 30, ai[0], FontTiny)
-            else:
-                dis.text(0, 27, ai[0], FontTiny)
-                dis.text(0, 27+7, ai[1], FontTiny)
-
-        dis.busy_bar(False)     # includes show
-
-    redraw()
-
-    from ux import ux_wait_keyup
-
-    while 1:
-        ch = await ux_wait_keyup()
-
-        if ch == '1':
-            invert = not invert
-            redraw()
-            continue
-        elif ch in 'xy':
-            return
-        if ch == '5' or ch == '7':
-            if idx > 0:
-                idx -= 1
-        elif ch == '8' or ch == '9':
-            if idx != len(addrs)-1:
-                idx += 1
-        
-        addr = addrs[idx]
-        data = render(addr)
-        redraw()
 
 # EOF
